+++ client/src/components/Trainercards.tsx
@@ -16,16 +16,26 @@ function TrainerCards({
   trainerfav3,
 }: TrainerInterface) {
   return (
-    <section className="trainer-card">
-      <h1>{trainername}</h1>
-      <img className="trainerimg" src={trainerimg} alt={trainername} />
-      <p>{trainerdescription}</p>
-      <section className="favpoke">
-        <img src={trainerfav1} alt={`Pokémon préféré de${trainername}`} />
-        <img src={trainerfav2} alt={`Pokémon préféré de${trainername}`} />
-        <img src={trainerfav3} alt={`Pokémon préféré de${trainername}`} />
-      </section>
-    </section>
+    <>
+      <article className="trainer-cards">
+
+        <section className="trainerInfos">
+          <div className="trainerPres">
+            <h2 className="trainerName">{trainername}</h2>
+            <img className="trainerImg" src={trainerimg} alt={trainername} />
+          </div>
+          <p className="trainerDescription">{trainerdescription}</p>   
+        </section>
+
+
+        <section className="favPoke">
+          <img src={trainerfav1} alt={`Pokémon préféré de${trainername}`} />
+          <img src={trainerfav2} alt={`Pokémon préféré de${trainername}`} />
+          <img src={trainerfav3} alt={`Pokémon préféré de${trainername}`} />
+        </section>
+
+      </article>
+    </>
   );
 }
