+++ client/src/main.tsx
@@ -9,7 +9,7 @@ import App from "./App";
 
 // Import additional components for new routes
 // Try creating these components in the "pages" folder
-import TrainerList from "./pages/TrainerList";
+import TrainerList from "./components/TrainerList";
 
 /* ************************************************************************* */
